// front/src/Models/AdminActionType.ts
export enum AdminActionType {
    ADD_APPOINTMENT = "addAppointment",
    ADD_APPOINTMENT_TO_PATIENT = "addAppointmentToPatient",
    ADD_PATIENT = "addPatient",
    ALL_APPOINTMENTS = "allAppointments",
    GET_APPOINTMENTS_BY_DOCTOR = "getAppointmentsByDoctor",
    GET_ALL_PATIENTS = "getAllPatients",
    DELETE_APPOINTMENT = "deleteAppointment",
    DELETE_PATIENTS = "deletePatients",
    GET_SINGLE_APPOINTMENT = "getSingleAppointment",
    GET_SINGLE_PATIENT = "getSinglePatient",
    UPDATE_APPOINTMENT = "updateAppointment",
    UPDATE_PATIENT = "updatePatient",
    GET_ALL_APPOINTMENTS_BY_DOCTOR_TYPE = "getAllAppointmentsByDoctorType"
}